Remove 'splitkeep'
I've had this off for so long.  It's too bad because I was very excited
about it and even donated a small amount to the guy who added it.  The
main reason I used it was for FZF but now with floating windows it's
mostly unnecessary.  There are still times it would be nice BUT the way
it moves the cursor if it's too close to the bottom of the screen is
brutal, so off you go.